create parental class for stubbing.
